начало перехода в компонент модального окна
