+++ frontend/src/main.js
@@ -1,18 +1,18 @@
 import './assets/styles/main.css'
 
 import { createApp } from 'vue'
-import { createPinia } from 'pinia'
+//import { createPinia } from 'pinia'
 import axios from "axios";
 import App from './App.vue'
 import router from './router'
 import themeMixin from './mixins/theme.js'
 
 const app = createApp(App)
-const pinia = createPinia()
+//const pinia = createPinia()
 
 app.mixin(themeMixin)
 app.use(router)
-app.use(pinia)
+//app.use(pinia)
 
 app.config.globalProperties.$axios = axios;
