+++ frontend/src/router/index.js
@@ -4,7 +4,6 @@ import authRouters from "./auth.js"
 import adminRouters from "./admin.js"
 import defaultRouters from "./default.js"
 import NotFoundView from "@/views/NotFoundView.vue";
-import {useUserStore} from "@/stores/user.js";
 
 const routes = [
         ...authRouters,
@@ -26,43 +25,43 @@ router.beforeEach(async(to, from, next) => {
     const isAuthenticated = tokenService.isAuthenticated()
     const isUsingTempTokens = tokenService.isUsingTempTokens()
 
-    if (to.meta.requiresActivationToken){
-        if(isAuthenticated){
-            next({name: 'Home'})
-            return
-        }
-        else{
-            if (!isUsingTempTokens){
-                next({name: 'Login'})
-                return
-            }
-        }
-    }
+    // if (to.meta.requiresActivationToken){
+    //     if(isAuthenticated){
+    //         next({name: 'Home'})
+    //         return
+    //     }
+    //     else{
+    //         if (!isUsingTempTokens){
+    //             next({name: 'Login'})
+    //             return
+    //         }
+    //     }
+    // }
+    //
+    // if (to.meta.requiresAuth){
+    //     if(!isAuthenticated) {
+    //         next({
+    //             name: 'Login',
+    //             query: {redirect: to.fullPath}
+    //         })
+    //         return
+    //     }
+    // }
+    //
+    // if (to.meta.requiresGuest) {
+    //     if (isAuthenticated) {
+    //         next({name: 'Home'})
+    //         return
+    //     }
+    // }
 
-    if (to.meta.requiresAuth){
-        if(!isAuthenticated) {
-            next({
-                name: 'Login',
-                query: {redirect: to.fullPath}
-            })
-            return
-        }
-    }
-
-    if (to.meta.requiresGuest) {
-        if (isAuthenticated) {
-            next({name: 'Home'})
-            return
-        }
-    }
-
-    if(to.meta.requiresAdmin){
-        const userStore = useUserStore()
-        if(!userStore.isAdmin){
-            // что-нибудь придумать для доступа
-            return
-        }
-    }
+    // if(to.meta.requiresAdmin){
+    //     const userStore = useUserStore()
+    //     if(!userStore.isAdmin){
+    //         // что-нибудь придумать для доступа
+    //         return
+    //     }
+    // }
 
     next()
 })
